requirements.txt
absl-py==2.1.0     
APScheduler==3.10.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
backports.shutil_which==3.5.2
beautifulsoup4==4.12.3       
bleak==0.21.1
bleak-winrt==1.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2    
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docutils==0.21.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0        
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.51.0
future==1.0.0
gast==0.4.0
gdx==3
godirect==1.1.4
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
hidapi==0.14.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.18.1
Janome==0.5.0
jedi==0.19.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivy-garden.matplotlib @ https://github.com/kivy-garden/matplotlib/archive/master.zip
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyGetWindow==0.0.9
Pygments==2.17.2
pyparsing==3.1.2
pypiwin32==223
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
PyRect==0.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-version==0.0.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wikipedia==1.4.0
wrapt==1.16.0
xarray==2024.3.0
zipp==3.18.1
pyautogui
mss
mysql-connector